publications/lion12/requirements.txt
altgraph==0.15
bleach==1.5.0
deap==1.2.2
decorator==4.2.1
future==0.16.0
graphviz==0.8.2
h5py==2.7.1
html5lib==0.9999999
hyperopt==0.1
Keras==2.1.2
lxml==3.5.0
macholib==1.9
Markdown==2.6.10
networkx==2.1
nose==1.3.7
numpy==1.13.0
pandas==0.19.2
pefile==2017.11.5
protobuf==3.5.0.post1
pydot==1.2.3
PyInstaller==3.3.1
pymongo==3.6.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.9.1
scipy==1.0.0
six==1.11.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
Werkzeug==0.13
